+++ app/views/login.pug
@@ -12,7 +12,6 @@ html(lang="en")
     //- meta(charset="UTF-8")
     //- meta(name="viewport", content="width=device-width, initial-scale=1.0")
     title Login
-    //- link(rel='stylesheet', href='../CSS/style.css')
     link(rel="stylesheet", type='text/css', href="/stylesheets/style.css")
     link(rel="stylesheet", type='text/css', href="/stylesheets/codegenerator.css")
   body
