+++ app/views/securityPages/security-layout.pug
@@ -8,7 +8,7 @@ html(lang="en")
       title Document
     link(rel="stylesheet", type='text/css', href="/stylesheets/style.css")
     link(rel="stylesheet", type='text/css', href="/stylesheets/codegenerator.css")
-    script(src="/modules/script.js")
+    
   body
     .dashboard_top
       .dashboard_top_left
@@ -30,4 +30,6 @@ html(lang="en")
 
       .main_body_div
         block content
+
+    script(src="/models/script.js")
